+++ src/config/database.js
@@ -1,5 +1,5 @@
 module.exports = {
-  development: { username: 'salamander', password: 'bnx3rhej33', database: 'salamanderdev', host: 'localhost', dialect: 'postgres' },
-  test: { username: 'salamander', password: 'bnx3rhej33', database: 'salamanderdev', host: 'localhost', dialect: 'postgres' },
-  production: { username: 'salamander', password: 'bnx3rhej33', database: 'salamanderdev', host: 'localhost', dialect: 'postgres' }
+  development: { username: 'salamander', password: 'salamander', database: 'salamanderdev', host: 'localhost', dialect: 'postgres' },
+  test: { username: 'salamander', password: 'salamander', database: 'salamanderdev', host: 'localhost', dialect: 'postgres' },
+  production: { username: 'salamander', password: 'salamander', database: 'salamanderdev', host: 'localhost', dialect: 'postgres' }
 }
