+++ src/infra/database/migrate/20190207214257-create-product-image.js
