+++ src/infra/database/models/product_image.js
